Table Template Service
Use it to only have one place for templates that are used in the table.
All templates regarding the table will be stored at the bottom of
the template of the "cd-table".

This service can be hooked on the columns input and the new
"useTemplates" output, to update all columns accordingly when the
templates references can be used.

Just call "overwriteSetColumns" to overwrite the "setColums" function
of this service and everything will work. The function that is passed
uses the same component scope it is called with as the second argument.

Signed-off-by: Stephan Müller <smueller@suse.com>