mgr/dashboard: CephFS directory component
Now it's possible to list directories of a CephFS inside the dashboard.
The behavior of showing sub directories is simple, yet intuitive. A click
on a directory will open the CephFS directory component beneath the
current table showing all the sub directories of the selected one.

Currently this table is not refreshed every 5 seconds as the listing of
directories is a heavier operation, but the user can hit the refresh
button anytime.

Fixes: https://tracker.ceph.com/issues/41575
Signed-off-by: Stephan Müller <smueller@suse.com>